+++ optaplanner-docs/src/main/asciidoc/QuickStart/PlainJava/CloudBalancingSolverConfiguration/CloudBalancingSolverConfiguration-section.adoc
@@ -18,7 +18,7 @@ The file is named `cloudBalancingSolverConfig.xml`.
 
   <!-- Score configuration -->
   <scoreDirectorFactory>
-    <easyScoreCalculatorClass>org.optaplanner.examples.cloudbalancing.optional.score.CloudBalancingEasyScoreCalculator</easyScoreCalculatorClass>
+    <easyScoreCalculatorClass>org.optaplanner.examples.cloudbalancing.optional.solver.score.CloudBalancingEasyScoreCalculator</easyScoreCalculatorClass>
     <!--<scoreDrl>org/optaplanner/examples/cloudbalancing/solver/cloudBalancingConstraints.drl</scoreDrl>-->
   </scoreDirectorFactory>
 
@@ -52,7 +52,7 @@ Further down, we will look into two alternatives to calculate the score: using a
 [source,xml,options="nowrap"]
 ----
   <scoreDirectorFactory>
-    <easyScoreCalculatorClass>org.optaplanner.examples.cloudbalancing.optional.score.CloudBalancingEasyScoreCalculator</easyScoreCalculatorClass>
+    <easyScoreCalculatorClass>org.optaplanner.examples.cloudbalancing.optional.solver.score.CloudBalancingEasyScoreCalculator</easyScoreCalculatorClass>
     <!--<scoreDrl>org/optaplanner/examples/cloudbalancing/solver/cloudBalancingConstraints.drl</scoreDrl>-->
   </scoreDirectorFactory>
 ----
