+++ optaplanner-examples/src/main/java/org/optaplanner/examples/cloudbalancing/domain/solver/CloudComputerStrengthComparator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.cloudbalancing.optional.domain;
+package org.optaplanner.examples.cloudbalancing.domain.solver;
 
 import static java.util.Comparator.comparing;
 import static java.util.Comparator.comparingInt;
