+++ optaplanner-examples/src/main/java/org/optaplanner/examples/cloudbalancing/optional/solver/move/CloudProcessSwapMoveFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,27 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.cloudbalancing.optional.move;
+package org.optaplanner.examples.cloudbalancing.optional.solver.move;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.ListIterator;
 
 import org.optaplanner.core.impl.heuristic.selector.move.factory.MoveListFactory;
 import org.optaplanner.examples.cloudbalancing.domain.CloudBalance;
-import org.optaplanner.examples.cloudbalancing.domain.CloudComputer;
 import org.optaplanner.examples.cloudbalancing.domain.CloudProcess;
 
-public class CloudComputerChangeMoveFactory implements MoveListFactory<CloudBalance> {
+public class CloudProcessSwapMoveFactory implements MoveListFactory<CloudBalance> {
 
     @Override
-    public List<CloudComputerChangeMove> createMoveList(CloudBalance cloudBalance) {
-        List<CloudComputerChangeMove> moveList = new ArrayList<>();
-        List<CloudComputer> cloudComputerList = cloudBalance.getComputerList();
-        for (CloudProcess cloudProcess : cloudBalance.getProcessList()) {
-            for (CloudComputer cloudComputer : cloudComputerList) {
-                moveList.add(new CloudComputerChangeMove(cloudProcess, cloudComputer));
+    public List<CloudProcessSwapMove> createMoveList(CloudBalance cloudBalance) {
+        List<CloudProcess> cloudProcessList = cloudBalance.getProcessList();
+        List<CloudProcessSwapMove> moveList = new ArrayList<>();
+        for (ListIterator<CloudProcess> leftIt = cloudProcessList.listIterator(); leftIt.hasNext();) {
+            CloudProcess leftCloudProcess = leftIt.next();
+            for (ListIterator<CloudProcess> rightIt = cloudProcessList.listIterator(leftIt.nextIndex()); rightIt.hasNext();) {
+                CloudProcess rightCloudProcess = rightIt.next();
+                moveList.add(new CloudProcessSwapMove(leftCloudProcess, rightCloudProcess));
             }
         }
         return moveList;
