+++ optaplanner-examples/src/main/java/org/optaplanner/examples/cloudbalancing/optional/solver/score/CloudBalancingEasyScoreCalculator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.cloudbalancing.optional.score;
+package org.optaplanner.examples.cloudbalancing.optional.solver.score;
 
 import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
 import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
