+++ optaplanner-examples/src/test/java/org/optaplanner/examples/cloudbalancing/app/CloudBalancingRealTimePlanningTurtleTest.java
@@ -24,10 +24,10 @@ import org.optaplanner.core.api.solver.ProblemFactChange;
 import org.optaplanner.examples.cloudbalancing.domain.CloudBalance;
 import org.optaplanner.examples.cloudbalancing.domain.CloudComputer;
 import org.optaplanner.examples.cloudbalancing.domain.CloudProcess;
-import org.optaplanner.examples.cloudbalancing.optional.realtime.AddComputerProblemFactChange;
-import org.optaplanner.examples.cloudbalancing.optional.realtime.AddProcessProblemFactChange;
-import org.optaplanner.examples.cloudbalancing.optional.realtime.DeleteComputerProblemFactChange;
-import org.optaplanner.examples.cloudbalancing.optional.realtime.DeleteProcessProblemFactChange;
+import org.optaplanner.examples.cloudbalancing.swingui.realtime.AddComputerProblemFactChange;
+import org.optaplanner.examples.cloudbalancing.swingui.realtime.AddProcessProblemFactChange;
+import org.optaplanner.examples.cloudbalancing.swingui.realtime.DeleteComputerProblemFactChange;
+import org.optaplanner.examples.cloudbalancing.swingui.realtime.DeleteProcessProblemFactChange;
 import org.optaplanner.examples.cloudbalancing.persistence.CloudBalancingGenerator;
 import org.optaplanner.examples.common.app.RealTimePlanningTurtleTest;
