+++ optaplanner-examples/src/test/java/org/optaplanner/examples/cloudbalancing/app/CloudBalancingSolveAllTurtleTest.java
@@ -18,7 +18,7 @@ package org.optaplanner.examples.cloudbalancing.app;
 
 import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
 import org.optaplanner.examples.cloudbalancing.domain.CloudBalance;
-import org.optaplanner.examples.cloudbalancing.optional.score.CloudBalancingMapBasedEasyScoreCalculator;
+import org.optaplanner.examples.cloudbalancing.optional.solver.score.CloudBalancingMapBasedEasyScoreCalculator;
 import org.optaplanner.examples.common.app.CommonApp;
 import org.optaplanner.examples.common.app.UnsolvedDirSolveAllTurtleTest;
