+++ optaplanner-examples/src/test/java/org/optaplanner/examples/cloudbalancing/optional/solver/score/CloudBalancingConstraintProviderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.cloudbalancing.optional.score;
+package org.optaplanner.examples.cloudbalancing.optional.solver.score;
 
 import org.junit.jupiter.api.Test;
 import org.optaplanner.examples.cloudbalancing.domain.CloudBalance;
